projects/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0
django-ajax-datatable==4.4.3
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-filter==21.1
django-jazzmin==2.5.0
django-mathfilters==1.0.0
django-tinymce==3.4.0
idna==3.3
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
